Reactivate accidentally deactivated tests
